begin restructuring of pipeline_components: moved overlayers into one directory, with metamap subdirectory
