+++ medacy/pipeline_components/__init__.py
@@ -1,9 +1,7 @@
 """Import pipeline components"""
-from medacy.pipeline_components.annotation.gold_annotator_component import GoldAnnotatorComponent
+from medacy.pipeline_components.feature_overlayers.gold_annotator_component import GoldAnnotatorComponent
 from medacy.pipeline_components.feature_extraction.discrete_feature_extractor import FeatureExtractor
 from medacy.pipeline_components.learners.bilstm_crf_learner import BiLstmCrfLearner
-from medacy.pipeline_components.metamap.metamap import MetaMap
-from medacy.pipeline_components.metamap.metamap_component import MetaMapComponent
 from medacy.pipeline_components.tokenization.character_tokenizer import CharacterTokenizer
 from medacy.pipeline_components.tokenization.clinical_tokenizer import ClinicalTokenizer
 from medacy.pipeline_components.tokenization.systematic_review_tokenizer import SystematicReviewTokenizer
