+++ medacy/pipeline_components/feature_overlayers/gold_annotator_component.py
