--- medacy/pipeline_components/patterns/__init__.py
@@ -1,1 +1,0 @@
-from .table_matcher_component import TableMatcherComponent
