+++ medacy/pipelines/clinical_pipeline.py
@@ -3,9 +3,9 @@ import spacy
 
 from medacy.pipelines.base import BasePipeline
 from medacy.pipeline_components.tokenization.clinical_tokenizer import ClinicalTokenizer
-from medacy.pipeline_components.annotation.gold_annotator_component import GoldAnnotatorComponent
-from medacy.pipeline_components.metamap.metamap_component import MetaMapComponent
-from medacy.pipeline_components.metamap.metamap import MetaMap
+from medacy.pipeline_components.feature_overlayers.gold_annotator_component import GoldAnnotatorComponent
+from medacy.pipeline_components.feature_overlayers.metamap import MetaMapComponent
+from medacy.pipeline_components.feature_overlayers.metamap import MetaMap
 from medacy.pipeline_components.feature_extraction.discrete_feature_extractor import FeatureExtractor
