+++ medacy/pipelines/lstm_systematic_review_pipeline.py
@@ -1,7 +1,7 @@
 import spacy
 
 from medacy.pipelines.base import BasePipeline
-from medacy.pipeline_components.annotation.gold_annotator_component import GoldAnnotatorComponent
+from medacy.pipeline_components.feature_overlayers.gold_annotator_component import GoldAnnotatorComponent
 from medacy.pipeline_components.feature_extraction.discrete_feature_extractor import FeatureExtractor
 from medacy.pipeline_components.learners.bilstm_crf_learner import BiLstmCrfLearner
 from medacy.pipeline_components.tokenization.systematic_review_tokenizer import SystematicReviewTokenizer
