+++ medacy/pipelines/scispacy_pipeline.py
@@ -3,10 +3,10 @@ import spacy
 from spacy.tokenizer import Tokenizer
 
 from medacy.pipelines.base import BasePipeline
-from medacy.pipeline_components.annotation.gold_annotator_component import GoldAnnotatorComponent
+from medacy.pipeline_components.feature_overlayers.gold_annotator_component import GoldAnnotatorComponent
 from medacy.pipeline_components.feature_extraction.discrete_feature_extractor import FeatureExtractor
-from medacy.pipeline_components.metamap.metamap import MetaMap
-from medacy.pipeline_components.metamap.metamap_component import MetaMapComponent
+from medacy.pipeline_components.feature_overlayers.metamap import MetaMap
+from medacy.pipeline_components.feature_overlayers.metamap import MetaMapComponent
 
 
 class ScispacyPipeline(BasePipeline):
