+++ medacy/pipelines/systematic_review_pipeline.py
@@ -2,10 +2,10 @@ import sklearn_crfsuite
 import spacy
 
 from medacy.pipelines.base.base_pipeline import BasePipeline
-from medacy.pipeline_components.annotation.gold_annotator_component import GoldAnnotatorComponent
+from medacy.pipeline_components.feature_overlayers.gold_annotator_component import GoldAnnotatorComponent
 from medacy.pipeline_components.feature_extraction.discrete_feature_extractor import FeatureExtractor
-from medacy.pipeline_components.metamap.metamap import MetaMap
-from medacy.pipeline_components.metamap.metamap_component import MetaMapComponent
+from medacy.pipeline_components.feature_overlayers.metamap import MetaMap
+from medacy.pipeline_components.feature_overlayers.metamap import MetaMapComponent
 from medacy.pipeline_components.tokenization.systematic_review_tokenizer import SystematicReviewTokenizer
